python-runner/pyproject.toml
[tool.poetry]
name = "brain-runner"
version = "0.1.0"
description = "Python runner for AI Stack"
authors = ["AI Assistant <ai@example.com>"]
readme = "README.md"
packages = [{include = "core"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = "^0.32.0"
psycopg2-binary = "^2.9.10"
requests = "^2.32.0"
openai = "^2.14.0"
qdrant-client = "^1.16.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
